requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
groq>=0.9.0
google-generativeai>=0.3.0
chromadb>=0.5.0
sentence-transformers>=2.7.0
PyPDF2>=3.0.1
pdfplumber>=0.10.3
python-docx>=1.1.0
rank-bm25>=0.2.2
numpy>=1.26.0
python-dotenv>=1.0.0
